+++ Sources/Models/ALKConfiguration.swift
@@ -148,6 +148,9 @@ public struct ALKConfiguration {
     // If true, Then message search will be enabled
     public var isMessageSearchEnabled: Bool = false
 
+    // If true, then in-app notification banner will be disabled
+    public var isInAppNotificationBannerDisabled: Bool = false
+
     /// If true, contact share option in chatbar will be hidden.
     @available(*, deprecated, message: "Use .chatBar.optionsToShow instead")
     public var hideContactInChatBar: Bool = false {