+++ CHANGELOG.md
@@ -8,7 +8,11 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 ### Enhancements
 - Added support for syncing package details when a suspension screen is shown.
 - Added an option to change message status(read receipt) style. Now icon for message status can be changed from outside or it can be hidden.
-
+- Added a config to disable in-app notification banner.
+</br>You can use below config to disable in-app notification banner:
+```
+	config.isInAppNotificationBannerDisabled = true
+```
 ## [5.0.0] - 2020-02-14
 
 ### Enhancements